dima/task2/requirements.txt
Django
django-environ
django-bootstrap-v5
djangorestframework
djangorestframework-simplejwt
psycopg2-binary>=2.9.0
psycopg2